+++ CHANGELOG.md
@@ -8,11 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.2.0] - 2022/08/05
 ### Added
 - Properly parse processing instruction nodes as XML wrappers.
-### Fixed
-- TODO:
+- Add `writeProcessingInstructions` and `minify` options to `XmlFormattingOptions`.
 ### Changed
 - Rename `includeProcessingInstructions` option in `XmlFormattingOptions` to `writeXmlDeclaration`.
-- Add `writeProcessingInstructions`option to `XmlFormattingOptions`.
 
 ## [0.1.3] - 2022/02/17
 ### Changed